Version piplite with the `bump-version` script

+++ scripts/bump-version.py
@@ -23,13 +23,15 @@ APP_JUPYTERLITE_JSON = ROOT / "app" / "jupyter-lite.json"
 
 PYOLITE_PACKAGE = ROOT / "packages" / "pyolite-kernel"
 PYOLITE_PACKAGE_JSON = PYOLITE_PACKAGE / "package.json"
-PYOLITE_KERNEL_SOURCE = PYOLITE_PACKAGE / "src" / "kernel.ts"
 PYOLITE_INIT_PY = PYOLITE_PACKAGE / "py" / "pyolite" / "pyolite" / "__init__.py"
+PIPLITE_INIT_PY = PYOLITE_PACKAGE / "py" / "piplite" / "piplite" / "__init__.py"
 
 
 def replace_in_file(path, pattern, replacement):
     source = path.read_text(**ENC)
     replaced = re.sub(pattern, replacement, source)
+    if replaced == source:
+        raise ValueError("pattern not found")
     path.write_text(replaced, **ENC)
 
 
@@ -43,17 +45,16 @@ def postbump():
 
     # bump pyolite wheel import
     replace_in_file(
-        PYOLITE_KERNEL_SOURCE,
-        "pyolite-(.*)-py3-none-any.whl",
-        f"pyolite-{py_version}-py3-none-any.whl",
+        PYOLITE_INIT_PY, "__version__ = (.*)", f'__version__ = "{py_version}"'
     )
     replace_in_file(
-        PYOLITE_INIT_PY, "__version__ = (.*)", f'__version__ = "{py_version}"'
+        PIPLITE_INIT_PY, "__version__ = (.*)", f'__version__ = "{py_version}"'
     )
 
     # bump pyolite version
     pyolite_json = json.loads(PYOLITE_PACKAGE_JSON.read_text(**ENC))
     pyolite_json["pyolite"]["packages"]["py/pyolite"] = py_version
+    pyolite_json["pyolite"]["packages"]["py/piplite"] = py_version
     PYOLITE_PACKAGE_JSON.write_text(json.dumps(pyolite_json), **ENC)
     run(f"yarn prettier --write {PYOLITE_PACKAGE_JSON}")
